==> notebooks/04_modelado.ipynb <==
#CELDA 1: Configuraciones básicas

import json
import pandas as pd
import numpy as np
import sys
import joblib

from pathlib import Path

from sklearn.svm import LinearSVC
from sklearn.neural_network import MLPClassifier
from sklearn.decomposition import TruncatedSVD
from sklearn.ensemble import VotingClassifier
from sklearn.pipeline import Pipeline

from datetime import datetime

%run 00_setup.ipynb

#RUTAS
from src.config.rutas import (
    OUTPUT_DIR,
    RUTA_DATASET,
    RANDOM_STATE
)

from sklearn.metrics import (
    accuracy_score,
    precision_score,
    recall_score,
    f1_score,
    roc_auc_score,
    classification_report,
    confusion_matrix
)

#Se importan las funciones y rutas creadas en los modulos .py de /feat_aux
from src.feat_aux import (
    preprocesar_dataset,
    dividir_train_test_estratificado,
    construir_tfidf,
    ajustar_y_vectorizar,
    calc_dispersion,
    guardar_meta
)

VERSION = "eva_hard_voting"

#CELDA 2: Carga del dataset

dt = pd.read_csv(RUTA_DATASET)

print("Num. de columnas y filas:")
print(dt.shape)
dt.head()

dt_preprocesado = preprocesar_dataset(dt)       #Se aplica internamente la funcion de limpiar
print("Forma tras preprocesar:", dt_preprocesado.shape)
dt_preprocesado.head()

#CELDA 3: Division estratificada train/test

dt_train, dt_test = dividir_train_test_estratificado(dt_preprocesado)

print("Train:", dt_train.shape)
print("Test:", dt_test.shape)

print("\nDistribucion por clases (train):")
print(dt_train["label"].value_counts(), "\n")
print(dt_train["label"].value_counts(normalize=True))

print("\nDistribucion por clases (test):")
print(dt_test["label"].value_counts(), "\n")
print(dt_test["label"].value_counts(normalize=True))


print("\nDistribucion por vector (train):")
print(dt_train["vector"].value_counts(), "\n")

print("\nDistribucion por vector (test):")
print(dt_test["vector"].value_counts(), "\n")


#CELDA 4: Contruccion de TF_IDF y posterior ajuste

tfidf = construir_tfidf(
    ngram_range=(1, 2),
    min_dt=5,
    max_dt=0.9,
    max_features=50000
)

x_train, x_test, y_train, y_test = ajustar_y_vectorizar(tfidf, dt_train, dt_test)

print("x_train:", x_train.shape)
print("x_test:", x_test.shape)
print("Etiquetas unicas:", np.unique(y_train))

guardar_meta(tfidf, dt_train, dt_test, OUTPUT_DIR, version=VERSION)

#CELDA 5: Dispersion de las matrices

disp_train = calc_dispersion(x_train)
disp_test = calc_dispersion(x_test)

print(f"Dispersion del train: {disp_train:.4f}")
print(f"Dispersion del test: {disp_test:.4f}")

#CELDA 6: Entrenamiento del modelo SVM:

svm = LinearSVC(
    class_weight="balanced",         #considera el desbalance que hay y penaliza los errores en las clases minoritarias
    random_state=RANDOM_STATE
)

#Se entrena el modelo
svm.fit(x_train, y_train)

#Se genera la prediccion
y_pred_xtest_svm = svm.predict(x_test)

#Metricas
accuracy_svm = accuracy_score(y_test, y_pred_xtest_svm)
precision_svm = precision_score(y_test, y_pred_xtest_svm, average="macro", zero_division=0)
recall_svm = recall_score(y_test, y_pred_xtest_svm, average="macro", zero_division=0)
f1_mac_svm = f1_score(y_test, y_pred_xtest_svm, average="macro", zero_division=0)       #Calcula f1 equitativamente cada clase
f1_wei_svm = f1_score(y_test, y_pred_xtest_svm, average="weighted", zero_division=0)    #Calcula f1 segun la frecuencia real

print(".: SVM :.")
print(f"Accuracy: {accuracy_svm:.4f}")
print(f"Precison: {precision_svm:.4f}")
print(f"Recall: {recall_svm:.4f}")
print(f"F1-macro: {f1_mac_svm:.4f}")
print(f"F1-weighted: {f1_wei_svm:.4f}")

#Se muestra un reporte conjunto
print("\nReport de clasificación SVM:\n")
print(classification_report(y_test, y_pred_xtest_svm, zero_division=0))

#Muestra una tabla de predicciones que fueron segun TP, TN, FP, FN
print("\nMatriz de confusion:\n")
print(confusion_matrix(y_test, y_pred_xtest_svm))


#CELDA 7: Entrenamiento del modelo MLP

#Se reduce la dimensionalidad sobre TF-IDF para no saturar el sistema al depender de mucha RAM. De esta manera es reproducible
svd = TruncatedSVD(n_components=300, random_state=RANDOM_STATE)

#Se transforma la matrix dispersa en completa
x_train_reducido = svd.fit_transform(x_train)
x_test_reducido = svd.transform(x_test)

mlp = MLPClassifier(
    hidden_layer_sizes=(128, 64),       #Define la estructura de la red neuronal siendo densamente conectada con 128 neuronas en la primera capa y 64 neuronas en la segunda
    activation="relu",                  #Activacion estandar para la mayoria de los modelos
    solver="adam",                      #Se pasa el optimizador mas recomendado para NLP
    random_state=RANDOM_STATE,
    max_iter=40,                        #Se define un maximo de 40 iteraciones para entrenar la red
    early_stopping=True,                #Activa la parada temprana
    n_iter_no_change=3,                 #Si en 3 iteraciones el modelo no mejora la perdida de validacion, se detiene automaticamente.
    verbose=True                        #Imprime la informacion durante el entrenamiento para visualizar resultados de las iteraciones
)

#Se entrena el modelo
mlp.fit(x_train_reducido, y_train)

#Se genera la prediccion
y_pred_xtest_mlp = mlp.predict(x_test_reducido)
y_prob_xtest_mlp = mlp.predict_proba(x_test_reducido) [:, 1]

#Metricas
accuracy_mlp = accuracy_score(y_test, y_pred_xtest_mlp)
precision_mlp = precision_score(y_test, y_pred_xtest_mlp, average="macro", zero_division=0)
recall_mlp = recall_score(y_test, y_pred_xtest_mlp, average="macro", zero_division=0)
f1_mac_mlp = f1_score(y_test, y_pred_xtest_mlp, average="macro", zero_division=0)       #Calcula f1 equitativamente cada clase
f1_wei_mlp = f1_score(y_test, y_pred_xtest_mlp, average="weighted", zero_division=0)    #Calcula f1 segun la frecuanecia real
auc_mlp = roc_auc_score(y_test, y_prob_xtest_mlp)

print("\n.: MLP :.")
print(f"Accuracy: {accuracy_mlp:.4f}")
print(f"Precison: {precision_mlp:.4f}")
print(f"Recall: {recall_mlp:.4f}")
print(f"F1-macro: {f1_mac_mlp:.4f}")
print(f"F1-weighted: {f1_wei_mlp:.4f}")
print(f"AUC-ROC: {auc_mlp:.4f}")

print("\nReport de clasificación MLP:\n")
print(classification_report(y_test, y_pred_xtest_mlp, zero_division=0))

print("\nMatriz de confusion:\n")
print(confusion_matrix(y_test, y_pred_xtest_mlp))


#CELDA 8: Hard voting

#Define una pipeline que encadena varios pasos, en este caso la reduccion de la dimensionalidad y el modelo MLP
mlp_pipeline = Pipeline([
    ("svd", TruncatedSVD(n_components=300, random_state=RANDOM_STATE)),
    ("mlp", MLPClassifier(
    hidden_layer_sizes=(128, 64),
    activation="relu",
    solver="adam",
    random_state=RANDOM_STATE,
    max_iter=40,
    early_stopping=True,
    n_iter_no_change=3,
    verbose=True
    ))
])

#Se crea el modelo en ensamblado que combina ambos modelos (SVM y MLP)
ens = VotingClassifier(
    estimators=[
        ("svm", svm),
        ("mlp", mlp_pipeline)
    ],
    voting="hard"
)

ens.fit(x_train, y_train)

y_pred_xtest_ens = ens.predict(x_test)

#Metricas
accuracy_ens = accuracy_score(y_test, y_pred_xtest_ens)
precision_ens = precision_score(y_test, y_pred_xtest_ens, average="macro", zero_division=0)
recall_ens = recall_score(y_test, y_pred_xtest_ens, average="macro", zero_division=0)
f1_mac_ens = f1_score(y_test, y_pred_xtest_ens, average="macro", zero_division=0)       #Calcula f1 equitativamente cada clase
f1_wei_ens = f1_score(y_test, y_pred_xtest_ens, average="weighted", zero_division=0)    #Calcula f1 segun la frecuanecia real

print("\n.: HARD VOTING :.")
print(f"Accuracy: {accuracy_ens:.4f}")
print(f"Precison: {precision_ens:.4f}")
print(f"Recall: {recall_ens:.4f}")
print(f"F1-macro: {f1_mac_ens:.4f}")
print(f"F1-weighted: {f1_wei_ens:.4f}")

print("\nReport de clasificación HARD VOTING:\n")
print(classification_report(y_test, y_pred_xtest_ens, zero_division=0))

print("\nMatriz de confusion:\n")
print(confusion_matrix(y_test, y_pred_xtest_ens))



#CELDA 9: Comparativa SVM vs MLP vs Ensemble

comparacion_modelos = pd.DataFrame({
    "Modelo": ["SVM", "MLP", "Ensemble (Hard Voting)"],
    "Accuracy": [float(accuracy_svm), float(accuracy_mlp), float(accuracy_ens)],
    "Precision": [float(precision_svm), float(precision_mlp), float(precision_ens)],
    "Recall": [float(recall_svm), float(recall_mlp), float(recall_ens)],
    "F1-macro": [float(f1_mac_svm), float(f1_mac_mlp), float(f1_mac_ens)],
    "F1-weighted": [float(f1_wei_svm), float(f1_wei_mlp), float(f1_wei_ens)],
})

cols_metricas = ["Accuracy", "Precision", "Recall", "F1-macro", "F1-weighted"]

comparacion_modelos.style.format(
    {col: "{:.4f}" for col in cols_metricas}
)

#CELDA 10: Guardado de los modelos finales

svm_path = OUTPUT_DIR / f"modelo_svm_{VERSION}.joblib"
mlp_path = OUTPUT_DIR / f"modelo_mlp_{VERSION}.joblib"
ens_path = OUTPUT_DIR / f"modelo_ens_{VERSION}.joblib"

joblib.dump(svm, svm_path)
joblib.dump(mlp, mlp_path)
joblib.dump(ens, ens_path)

print("Modelos guardados:")
print("SVM:", svm_path)
print("MLP:", svm_path)
print("Ensamble:", ens_path)

#CELDA 11: Guardar metricas en JSON

metricas = {
        "version": VERSION,
        "generated_at": datetime.now().isoformat(timespec="seconds"),
        "random_state": RANDOM_STATE,
        "modelos": {
            "SVM": {
                "accuracy": float(accuracy_svm),
                "precision": float(precision_svm),
                "recall": float(recall_svm),
                "F1-macro": float(f1_mac_svm),
                "F1-weighted": float(f1_wei_svm),
                "model_path": str(svm_path)
            },
            "MLP": {
                "accuracy": float(accuracy_mlp),
                "precision": float(precision_mlp),
                "recall": float(recall_mlp),
                "F1-macro": float(f1_mac_mlp),
                "F1-weighted": float(f1_wei_mlp),
                "model_path": str(mlp_path)
            },
            "Ensamble": {
                "accuracy": float(accuracy_ens),
                "precision": float(precision_ens),
                "recall": float(recall_ens),
                "F1-macro": float(f1_mac_ens),
                "F1-weighted": float(f1_wei_ens),
                "model_path": str(ens_path)
            }
        },
        "dispersion": {
            "train": float(disp_train),
            "test": float(disp_test)
        }
}

#Guarda el archivo JSON con todas las metricas al abrirlo en modo escritura "w"
metricas_path = OUTPUT_DIR / f"metricas_modelos_{VERSION}.json"
with open(metricas_path, "w", encoding="utf-8") as f:
    json.dump(metricas, f, indent=4, ensure_ascii=False)

print("Metricas guardadas en:", metricas_path)
